requirements.txt
pandas
openpyxl
pyarrow
spacy
sentence-transformers
scikit-learn>=1.3.0
streamlit
plotly
openai
python-dotenv
